Update leds in handlers
