+++ Esp32Light/src/main.hpp
@@ -1,6 +1,6 @@
 #include <Arduino.h>
 #include "Globals.hpp"
-#include "LedUtils.hpp"
+#include "Utils.hpp"
 #include "SupportStructures.hpp"
 #include "LightToSerialParser.hpp"
 #include "LightController.hpp"
@@ -10,11 +10,6 @@ static CRGB defaultColorRIGHT = CRGB(0, 125, 255);
 
 void init_controllers();
 void init_lasers();
-void fadeLight();
-void fadeFlashLight();
-void cacheLeftColor(uint8_t r, uint8_t g, uint8_t b);
-void cacheRightColor(uint8_t r, uint8_t g, uint8_t b);
 void controlLight();
 void handleEvent();
-void ledwalkleft(t_laser *);
-void ledwalkright(t_laser *);
+void applyEventToGroup();
